loop through indexed list in defined order
in order to first install LibreOffice5 dependencies:
- libreoffice5.2-ure_5.2.2.2-2_amd64.deb
- libobasis5.2-core_5.2.2.2-2_amd64.deb
- libobasis5.2-librelogo_5.2.2.2-2_amd64.deb

add them as first items in the `_indexed_items` list, then continue
looping through ALL items of the list.

This works ok only in `apt` module or other idempotent modules

WARNING! Do not try this with modules that are not idempotent, it may
break your system.

Resolves: #30
Related: ansible/ansible-modules-core#1178
Signed-off-by: Daniel Andrei Minca <mandrei17@gmail.com>