remove obsolete ubuntu code
The templates and lines of code removed were from the past when the
configuration was destined for Debian systems.

After switching to ArchLinux, this code is now obsolete as it no longer
applies to current System (different configuration, package manager,
system files et. al)

As a result, these files are removed and no longer needed.

Resolves:
Related:
Signed-off-by: Daniel Andrei Minca <mandrei17@gmail.com>