add failsafe for autoenabling tmux
If tmux wasn't installed on a certain system, Admin wasn't able to SSH
in the host, thus blocking any TTY operations.

As a result, check if tmux is installed on a system and then load the
tmux configuration, otherwise don't load it and let the Admin access the
Host without any TTY errors.

Resolves:
Related:
Signed-off-by: Daniel Andrei Minca <mandrei17@gmail.com>