disable recursive key pressing in tmux
After switching pane to any direction, it was hard to start typing
anything that included the 'pane-switching' key from tmux as it was
conflicting thus jumping to another pane while typing unix commands in
the shell.

As a result, to avoid conflicting with tmux keybindings, disable the
`-r` flag when jumping between panes.

Realised this in the comment below my StackOverflow answer:
https://unix.stackexchange.com/a/264052/50394

Resolves:
Related:
Signed-off-by: Daniel Andrei Minca <mandrei17@gmail.com>