fix the IP Addrs from Tmux config file
The IP Addresses along with Interface names were not displayed properly
as no `ifconfig` utility is installed on the system (since it's
deprecated for quite some time).

As a result, use the `ip` utility and parse the output with AWK in order
to cleanly format Interface Name and IPAddress in the bar.

Resolves:
Related:
Signed-off-by: Daniel Andrei Minca <mandrei17@gmail.com>