install dpaste for pasting code on the web easily
dpaste API is quite easy to understand, plus the website is very
minimalist:

http://dpaste.com/api/v2/

This allows us to paste code of logs and other stuff on the web easily,
in order to provide links on IRC or Telegram to help the Community easily
debug issues on our system. A simple paste could be achieved via the
command:

  cat ~/.ssh/config | dpaste -t Title -p UserBar -s cfg -e 1

This just says: posting the ~/.ssh/config file with title Title,
user UserBar, filetype cfg and expiry time 1 day.

Signed-off-by: Daniel Andrei Minca <mandrei17@gmail.com>